+++ Sensor/Core/SpiPort/ICaptorPort.cs
@@ -2,5 +2,5 @@ namespace Core.SpiPort;
 
 public interface ICaptorPort
 {
-    Task<int> GetTemperature();
+    Task<sbyte> GetTemperature();
 }
